Changed test scenario to account for the user opting to type in a number
Changed the rule-entry test scenario to first select that a number will be
chosen and then select that the number will be typed in (instead of just
selecting that a number will be chosen and assuming it will be typed in).
This is to account for the fact that other forms of selecting a number,
such as the result of some calculation or the value of some attribute, may
be used instead of a plain keyboard-typed value.