Added Vue components for selecting a value, currently only numbers
Added Vue component for selecting what type of value will be used, which
delegates the responsibility of actually selecting that value. Currently
only number-type values are supported.
Added Vue component for choosing a way to provide a number value, which
delegates the responsibility of yielding the exact value to a dynamically
chosen component.