Updated rule test to correctly select values in Vue Select components
Created a function to correctly select values in Vue Select components;
this is necessary because the component comes from an external library
and thus cannot be modified as easy as an HTML select. This function has
been extracted into an external file so it can be used in future tests.
Additionally, changed the rules-test to properly await for selections to
finish so that slotted components are updated correctly.