Changed rule-related Vue components to use actually modify the store
Added props and events to all Vue components in the rule-specification
hierarchy so that the existing state in the store is correctly displayed
and so that user actions successfully cascade upwards and modify that
state. Additionally, added the requisite mutations to the store to enable
such modification of state.

Also created JavaScript classes representing the yet-unrepresented objects
in the hierarchy, such as effects and selectors, in order to use them as
props.