Changed the signature of the 'triggering' methods for effects.
Effects now have an execute() method that takes as arguments the game state
(like the gamePlay subsection of the Vuex store state), a dispatch method
(with which they can call Vuex actions to make changes) and the
parameters (if any) associated with the triggering event.

Attribute selectors now have a getAttribute() method that takes as
arguments the game state and the trigger parameters and returns the name
of the attribute the selector is pointing to and a function that, when
given the store state as argument, will extract the object holding the
selector's attribute. (This complication is necessary because the
modification itself must happen in the store.) Attribute selectors also
have a getAttributeType() method now, a utility that returns the type of
attribute the selector is pointing to.

Valuie selectors now have a getValue() method that also takes as arguments
the game state and the trigger parameters and returns the current value
of the element the selector points to in the given state.