Added actions (and mutations) for the event-triggering mechanism.
Added the trigger action that will be called internally in order to launch
the execution of game effects.
Added an action to be dispatched when the game starts (an example of action
dispatched when a game event occurs, which then dispatches the appropriate
trigger action).
Added an action to be dispatched from within effects in order to change
the value of an attribute, plus the associated mutation.