Added Vue components for conditions; Added Vue component for if ... then.
Added components for specifying conditions in the user interface.
Added effect component for specifying that if a certain condition is met,
a certain effect should be executed.

Also modified the base effect component to stop mouse-click propagation,
since it contains a v-select and might, itself, be recursively contained
in another v-select.