CSS changes to make V-Selects grow less vertically.
Removed top margins and paddings across various V-Select components in an
effort to make it thinner so that complex don't grow so tall. Perhaps it is
a little much currently, though?
Also fixed an error in GameSpecAttributeListItem where it was using id
instead of class and updated it to use the global 'display-row' style.