Removed the 'Game' prefix from the exported name of JavaScript classes.
The prefix was originally added to prevent potential name clashes with
other objects (especially because HTML also has the concept of attributes).
However, no such conflicts have occurred nor seem likely to occur, while
the difference between the exported name and the file name / the name with
which the classes are imported in all other components is more likely to
produce confusion.