Added a uniquely identifying id to pieces.
This is because there was currently no way to tell if a piece that had been
moved was still the same piece or a different instance.
Currently, I did not change the way the selector for the player-selected
piece works. However, I have added a TODO item for how it could be changed
if issues arise.