Added reminder to check how licensing should be managed in the app.
Since we are using Vue Select, which is licensed by Jeff Sagal under the
MIT License, it is possible that the usage of his component should be
accompanied by some kind of notice indicating such. Since I do not know
currently what a such notice should include, I have attached a FIXME so the
issue is not forgotten.