Changed gameplay state to use 'Column' instead of 'Col' for tiles.
Previously, the tile attribute that identified the tile's column was 'Col'.
Since this was found to be less intuitive than 'Column', it was changed.

Also altered the piece cration mutation to re-fetch the target tile from
the live gameplay grid (based on the target tile's coordinates), just in
case reactivity issues might arise.