Added 'valueExpected' prop to attribute selectors.
The role of this prop is to allow parent components to specify the types of
attributes allowed. The components will automatically filter to only show
permitted-type attributes.
Also refactored computed option attributeOptions() in the from-attribute
component by breaking it down into first selecting the attribtues and then
computing the options from that.