Added JavaScript and Vue components for effect '... wins the game'.
Added components for the effect. Added new effect to the generic
GameSpecEffect parent. Added a new store action for ending the game with
a given winner.
Currently, the action simply alerts a message and then resets the game;
future implementations should aim to be more comprehensive in presenting
the victory.