Added CSS to prevent dropdown menu in selectors from crumpling up.
Because our selectors do not have a set width, the menu options collapse
to their minimum width, 160px. The changes submitted prevent this by
forcing the menu options to always be at least as wide as their content
and, by flex, at least as wide as the widest of them.